@@ -1,11 +1,18 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { UntypedFormControl, Validators } from '@angular/forms';
+import { UntypedFormControl, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { Subscription } from 'rxjs';
+import { NgIf, DecimalPipe } from '@angular/common';
+import { MatInput } from '@angular/material/input';
+import { MatFormField, MatLabel, MatSuffix, MatError } from '@angular/material/form-field';
+import { FlexModule } from '@ngbracket/ngx-layout/flex';
+import { MatCard } from '@angular/material/card';
 
 @Component({
-  selector: 'app-can-i-hit',
-  templateUrl: './can-i-hit.component.html',
-  styleUrls: ['./can-i-hit.component.scss']
+    selector: 'app-can-i-hit',
+    templateUrl: './can-i-hit.component.html',
+    styleUrls: ['./can-i-hit.component.scss'],
+    standalone: true,
+    imports: [MatCard, FlexModule, MatFormField, MatLabel, MatInput, FormsModule, ReactiveFormsModule, MatSuffix, NgIf, MatError, DecimalPipe]
 })
 export class CanIHitComponent implements OnInit, OnDestroy {
 
