@@ -1,17 +1,26 @@
 import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { MatMenuTrigger } from '@angular/material/menu';
+import { MatMenuTrigger, MatMenu, MatMenuItem } from '@angular/material/menu';
 import { DomSanitizer } from '@angular/platform-browser';
 import { CampaignDataService } from 'src/app/data/services/campaigns/campaign-data.service';
 import { EncounterDataService } from 'src/app/data/services/encounters/encounter-data.service';
 import { DeleteCampaignDialogComponent } from '../delete-campaign-dialog/delete-campaign-dialog.component';
 import { Campaign } from '../models/campaign';
 import { NewCampaignComponent } from '../new-campaign/new-campaign.component';
+import { RouterLink } from '@angular/router';
+import { MatDivider } from '@angular/material/divider';
+import { MatIcon } from '@angular/material/icon';
+import { MatList, MatListItem, MatListItemAvatar, MatListItemTitle, MatListItemLine } from '@angular/material/list';
+import { NgIf, NgFor, DatePipe } from '@angular/common';
+import { MatButton } from '@angular/material/button';
+import { MatCard } from '@angular/material/card';
 
 @Component({
-  selector: 'app-campaigns-landing',
-  templateUrl: './campaigns-landing.component.html',
-  styleUrls: ['./campaigns-landing.component.scss']
+    selector: 'app-campaigns-landing',
+    templateUrl: './campaigns-landing.component.html',
+    styleUrls: ['./campaigns-landing.component.scss'],
+    standalone: true,
+    imports: [MatCard, MatButton, NgIf, MatList, NgFor, MatListItem, MatMenuTrigger, MatListItemAvatar, MatIcon, MatListItemTitle, MatListItemLine, MatDivider, MatMenu, MatMenuItem, RouterLink, DatePipe]
 })
 export class CampaignsLandingComponent implements OnInit, OnDestroy {
   public campaigns: Campaign[] = [];
