@@ -1,11 +1,19 @@
 import { Component, Inject } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatDialogRef, MAT_DIALOG_DATA, MatDialogTitle, MatDialogContent, MatDialogActions, MatDialogClose } from '@angular/material/dialog';
 import { Campaign } from '../models/campaign';
+import { MatButton } from '@angular/material/button';
+import { MatDivider } from '@angular/material/divider';
+import { NgIf } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { MatInput } from '@angular/material/input';
+import { MatFormField, MatLabel } from '@angular/material/form-field';
 
 @Component({
-  selector: 'app-new-campaign',
-  templateUrl: './new-campaign.component.html',
-  styleUrls: ['./new-campaign.component.scss']
+    selector: 'app-new-campaign',
+    templateUrl: './new-campaign.component.html',
+    styleUrls: ['./new-campaign.component.scss'],
+    standalone: true,
+    imports: [MatDialogTitle, MatDialogContent, MatFormField, MatLabel, MatInput, FormsModule, NgIf, MatDivider, MatDialogActions, MatButton, MatDialogClose]
 })
 export class NewCampaignComponent {
   public readonly isEditingCurrentCampaign: boolean;
