@@ -8,8 +8,8 @@ describe('PrivacyPolicyComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ PrivacyPolicyComponent ]
-    })
+    imports: [PrivacyPolicyComponent]
+})
     .compileComponents();
   });
 
