@@ -8,8 +8,8 @@ describe('TermsComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ TermsComponent ]
-    })
+    imports: [TermsComponent]
+})
     .compileComponents();
   });
 
