@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { EncounterGeneratorComponent } from './components/encounter-generator/encounter-generator.component';
 import { MyEncountersComponent } from './components/my-encounters/my-encounters.component';
 
-const routes: Routes = [
+export const routes: Routes = [
   {
     path: '',
     component: EncounterGeneratorComponent
@@ -13,9 +13,3 @@ const routes: Routes = [
     component: MyEncountersComponent
   }
 ];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class EncountersRoutingModule { }
