@@ -8,8 +8,8 @@ describe('FeedbackComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ FeedbackComponent ]
-    })
+    imports: [FeedbackComponent]
+})
     .compileComponents();
   });
 
