@@ -1,15 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
+import { MatCard } from '@angular/material/card';
 
 @Component({
-  selector: 'app-feedback',
-  templateUrl: './feedback.component.html',
-  styleUrls: ['./feedback.component.scss']
+    selector: 'app-feedback',
+    templateUrl: './feedback.component.html',
+    styleUrls: ['./feedback.component.scss'],
+    standalone: true,
+    imports: [MatCard]
 })
-export class FeedbackComponent implements OnInit {
+export class FeedbackComponent {
 
   constructor() { }
-
-  ngOnInit(): void {
-  }
 
 }
